code function search and fix batch delete

+++ Project3_Member_Management_SpringBoot/src/main/java/com/project3/Member_Management_SpringBoot/controller/DeviceController.java
@@ -103,10 +103,12 @@ public class DeviceController {
             return "redirect:/admin/device-management?err=" + err;
         }
         List<Device> devices = deviceService.findAllDevicesLikeId(Integer.parseInt(idString));
-
+        if (devices == null || devices.isEmpty()) {
+            String err = "No devices found with the specified ID and year!";
+            return "redirect:/admin/device-management?err=" + err;
+        }
         for (Device device : devices) {
             deviceService.deleteDeviceById(device.getId());
-
         }
         String success = "Batch delete devices successfully";
         return "redirect:/admin/device-management?success=" + success;
